refactor(frontend): clean up page and settings panel components
- Removed unused imports and consolidated component imports for better readability.
- Introduced a new utility function for stage background styles.
- Added logic to conditionally display the admin tab in the SettingsPanel based on environment variables.
- Implemented a useEffect hook to manage active tab state in SettingsPanel.
- Enhanced VoiceInterface to include a loading phase state for improved UI feedback during voice interactions.
